requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.0
fastapi==0.110.0
filelock==3.12.4
h11==0.14.0
httptools==0.6.1
idna==3.6
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.4.0
jupyter_core==5.4.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==25.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.36.3
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.24.5
watchfiles==0.21.0
wcwidth==0.2.8
websockets==12.0
